Add custome design for exit and flag
