+++ .babelrc
@@ -1,3 +1,3 @@
 {
-  "presets": ["babel-preset-react", "babel-preset-env"],
+  "presets": ["@babel/preset-env", "@babel/preset-react"],
 }
