+++ client/src/__tests__/basic.test.js
@@ -1,6 +1,10 @@
 import React from 'react';
-import { shallow } from 'enzyme';
-import App from '../index.jsx';
+import Enzyme, { shallow } from 'enzyme';
+import Adapter from 'enzyme-adapter-react-16';
+import toJson from 'enzyme-to-json';
+import App from '../components/App.jsx';
+
+Enzyme.configure({ adapter: new Adapter });
 
 describe('Inital test', () => {
   // it('should render one <Search /> component', () => {
@@ -9,7 +13,7 @@ describe('Inital test', () => {
   // });
   test('Should pass a basic test', () => {
     const wrapper = shallow(<App />);
-    expect(wrapper).toHaveProperty('constructor');
+    expect(toJson(wrapper)).toHaveProperty('constructor');
   });
   // test('should render app', () => {
   //   shallow(<App />).contains(<div className="username">Val</div>);
